Make only a single REST request for each click in the UI
Before this change each click on a tree node in the Website
editor caused multiple REST requests to the backend. With this
change we bundle all requests into one operation performed on
the backend.

283555
Remove library files which should be ignored
Fixed get-started link bug when not on index page